+++ include/akushon/action.hpp
@@ -21,11 +21,11 @@
 #ifndef AKUSHON__ACTION_HPP_
 #define AKUSHON__ACTION_HPP_
 
-#include <akushon/motion.hpp>
 #include <akushon/pose.hpp>
 
-#include <map>
 #include <memory>
+#include <string>
+#include <vector>
 
 namespace akushon
 {
@@ -33,12 +33,23 @@ namespace akushon
 class Action
 {
 public:
-  void insert_motion(uint8_t id, std::shared_ptr<Motion> motion);
-  void delete_motion(uint8_t id);
+  explicit Action(std::string pose_name);
+
+  void insert_pose(Pose pose);
+  void insert_pose(uint8_t id, Pose pose);
+  void delete_pose(uint8_t id);
+
+  void set_name(std::string pose_name);
+  std::string get_name();
 
 private:
-  std::map<uint8_t, std::shared_ptr<Motion>> motion_list;
+  std::string name;
+
+  uint8_t next_motion_id;
+
+  static std::vector<Pose> poses;
 };
+
 }  // namespace akushon
 
 #endif  // AKUSHON__ACTION_HPP_
